Remove tree from programs
It sometimes gets shown on setup day. And I think it was in some
instructions once. Would be nice if the students have it.

But it is causing a call to brew that would otherwise not have happend
to many times. Which slows stuff down. And breaks if brew is broken.